Cleaned up the event handler

+++ View.h
@@ -5,7 +5,7 @@ struct View {
 	View(Uint16 _x, Uint16 _y, Uint16 _w, Uint16 _h, Uint16 _zoom)
 		: x(_x), y(_y), w(_w), h(_h), zoom(_zoom) {};
 
-	Uint16 x, y;
+	float x, y;
 	Uint16 w, h;
 
 	Uint16 zoom;
